refactor: move and rename add-or-update-address-modal
