+++ client-app/ui-kit/components/organisms/index.ts
@@ -1,6 +1,5 @@
 // FIXME: Remove private exports
 export { default as TableStatusBadge } from "./table/_internal/table-status-badge/table-status-badge.vue";
-export { default as VcAddOrUpdateAddressModal } from "./add-or-update-address-modal/vc-add-or-update-address-modal.vue";
 export { default as VcAddToCart } from "./add-to-cart/vc-add-to-cart.vue";
 export { default as VcConfirmationDialog } from "./confirmation-dialog/vc-confirmation-dialog.vue";
 export { default as VcPopup } from "./popup/vc-popup.vue";
