+++ client-app/ui-kit/components/organisms/types.d.ts
@@ -4,7 +4,6 @@ declare module "@vue/runtime-core" {
   // Glboal components is already declated interface which we want to augment
   // eslint-disable-next-line @typescript-eslint/naming-convention
   export interface GlobalComponents {
-    VcAddOrUpdateAddressModal: typeof Components.VcAddOrUpdateAddressModal;
     VcAddToCart: typeof Components.VcAddToCart;
     VcConfirmationDialog: typeof Components.VcConfirmationDialog;
     VcPopup: typeof Components.VcPopup;
